Populate a WorkContext in the created scope
Also setting Language, Currency, TaxDisplayType properties on it.
This is normally done by the WorkContextMiddleWare, but that is
not running in here.

If these WorkContext properties are not properly set it will cause issues
in the Liquid template items - for example:

Getting an IPaymentMethod.FriendyName normally uses ITranslationService,
but it will return empty string if the injected WorkContext does not have
a WorkingLanguage